Bring #2149 to post .NET 10

+++ crates/bindings-csharp/Runtime/Http.cs
@@ -331,9 +331,9 @@ public sealed class HttpClient
 
             var status = FFI.procedure_http_request(
                 requestBytes,
-                (uint)requestBytes.Length,
+                requestBytes.Length,
                 bodyBytes,
-                (uint)bodyBytes.Length,
+                bodyBytes.Length,
                 out var out_
             );
